[REFACTOR/#52] core > enum으로 변경
