+++ app/src/main/java/com/napzak/market/core/type/BottomSheetType.kt
@@ -1,0 +1,6 @@
+package com.napzak.market.core.type
+
+enum class BottomSheetType {
+    SORT,
+    GENRE_SEARCHING,
+}
