+++ app/src/main/java/com/napzak/market/presentation/explore/ExploreViewModel.kt
@@ -3,15 +3,15 @@ package com.napzak.market.presentation.explore
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.napzak.market.core.common.state.UiState
+import com.napzak.market.core.type.BottomSheetType
 import com.napzak.market.core.type.SortType
+import com.napzak.market.core.type.TradeType
 import com.napzak.market.domain.explore.model.ProductItem
 import com.napzak.market.domain.genre.model.Genre
 import com.napzak.market.presentation.explore.state.ExploreBottomSheetState
 import com.napzak.market.presentation.explore.state.ExploreProductInformation
 import com.napzak.market.presentation.explore.state.ExploreUiState
-import com.napzak.market.presentation.explore.type.ExploreBottomSheetType
 import com.napzak.market.presentation.explore.type.ExploreScreenType
-import com.napzak.market.presentation.explore.type.TradeType
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -275,15 +275,15 @@ class ExploreViewModel @Inject constructor(
         /* TODO: 좋아요 API 연결 및 기능 연결 */
     }
 
-    fun updateBottomSheetVisibility(type: ExploreBottomSheetType) {
+    fun updateBottomSheetVisibility(type: BottomSheetType) {
         when (type) {
-            ExploreBottomSheetType.SORT -> {
+            BottomSheetType.SORT -> {
                 _bottomSheetState.update {
                     it.copy(isSortBottomSheetVisible = !_bottomSheetState.value.isSortBottomSheetVisible)
                 }
             }
 
-            ExploreBottomSheetType.GENRE_SEARCHING -> {
+            BottomSheetType.GENRE_SEARCHING -> {
                 _bottomSheetState.update {
                     it.copy(isGenreSearchingBottomSheetVisible = !_bottomSheetState.value.isGenreSearchingBottomSheetVisible)
                 }
