+++ app/src/main/java/com/napzak/market/presentation/marketinfo/MarketInfoRoute.kt
@@ -31,7 +31,7 @@ import com.napzak.market.core.designsystem.theme.NapzakMarketTheme
 import com.napzak.market.core.type.SortType
 import com.napzak.market.core.type.MarketTab
 import com.napzak.market.domain.genre.model.Genre
-import com.napzak.market.presentation.explore.type.ExploreBottomSheetType
+import com.napzak.market.core.type.BottomSheetType
 import com.napzak.market.presentation.marketinfo.component.MarketFilterGroup
 import com.napzak.market.presentation.marketinfo.component.MarketInfoBottomSheetScreen
 import com.napzak.market.presentation.marketinfo.component.MarketInfoTopSection
@@ -69,17 +69,17 @@ fun MarketInfoRoute(
         onTradeTypeClick = viewModel::updateMarketTab,
         onGenreListClick = {
             viewModel.initGenreList()
-            viewModel.updateBottomSheetVisibility(ExploreBottomSheetType.GENRE_SEARCHING)
+            viewModel.updateBottomSheetVisibility(BottomSheetType.GENRE_SEARCHING)
         },
         onSoldOutClick = viewModel::updateSale,
         onUnopenClick = viewModel::updateUnopen,
-        onSortButtonClick = { viewModel.updateBottomSheetVisibility(ExploreBottomSheetType.SORT) },
+        onSortButtonClick = { viewModel.updateBottomSheetVisibility(BottomSheetType.SORT) },
         onItemClick = onProductDetailNavigate,
         onLikeClick = viewModel::updateItemLikeButton,
         onDismissRequest = viewModel::updateBottomSheetVisibility,
         onSortItemClick = {
             viewModel.updateSortType(it)
-            viewModel.updateBottomSheetVisibility(ExploreBottomSheetType.SORT)
+            viewModel.updateBottomSheetVisibility(BottomSheetType.SORT)
         },
         onTextChange = viewModel::changeSearchText,
         onGenreSelectButtonClick = viewModel::updateSelectedGenreList,
@@ -98,7 +98,7 @@ fun MarketInfoScreen(
     onSortButtonClick: () -> Unit,
     onItemClick: (Long) -> Unit,
     onLikeClick: (Long) -> Unit,
-    onDismissRequest: (ExploreBottomSheetType) -> Unit,
+    onDismissRequest: (BottomSheetType) -> Unit,
     onSortItemClick: (SortType) -> Unit,
     onTextChange: (String) -> Unit,
     onGenreSelectButtonClick: (List<Genre>) -> Unit,
@@ -164,7 +164,7 @@ fun MarketInfoSuccessScreen(
     onSortButtonClick: () -> Unit,
     onItemClick: (Long) -> Unit,
     onLikeClick: (Long) -> Unit,
-    onDismissRequest: (ExploreBottomSheetType) -> Unit,
+    onDismissRequest: (BottomSheetType) -> Unit,
     onSortItemClick: (SortType) -> Unit,
     onTextChange: (String) -> Unit,
     onGenreSelectButtonClick: (List<Genre>) -> Unit,
